+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 - Observable errors are now re-thrown in render, so that you can catch them with an Error Boundary
+- withObservable no longer has a dependency on RxJS package, which makes it leaner, more performant
+   and should hopefully fix persistent issues with rxjs-compat that some users have
 
 ## 1.0.8 - 2020-12-17